Added UI with basic authentication + base models and database
The structures are not yet final and have not been tried out. Implementatio is mostly skeletal and only has one simple view table that shows pending questions. There are seeders for languages and topics. This should not yet be used as the database will get many chanes before becoming stable.