References #3, simple emial notifications command
Created a Mail class for sending daily notifications to Language Expert users. Also created a command that will send out notifications, provided that all conditions are met. Current conditions are: there are PendingQuestions with status set to pending, that have not been changed yet and having been created yesterday. Also added a langage relation to users with NULL value by default. UI does not yet allow for the language to be set. An extension to user management is required for that. Language option will be considered in case a role assumes language choice being present (example: Language Expert).