Updated README file.
Added initial installation and development instructions. These are incomplete and an administrator creation command is missing (or at least a user account creation command) as the registration is disabled by default and it does not make any sense to change the configuration to temporarily enable it in order to create the initial administrator account.