Removed debounce from all inputs as it has issues with trim
Debounce can not be used with trim property of inputs as trim is applied during the data being set which has an effect of an unexpected removal of trailing space. If debounce would need to be used, then trim should be removed or applied after debounced event/action (example: update) is triggered. Current solution was to remove it altogether.